backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
google-generativeai>=0.8.0
chromadb>=0.4.22
pymupdf>=1.23.0
python-docx>=1.1.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
ollama>=0.4.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
torch>=2.1.0
sentencepiece>=0.1.99
